+++ .php_cs.dist
@@ -1,0 +1,18 @@
+<?php
+
+$finder = PhpCsFixer\Finder::create()
+    ->in([
+        __DIR__ . '/src',
+        __DIR__ . '/tests',
+    ])
+;
+
+return PhpCsFixer\Config::create()
+    ->setRules([
+        '@PSR1' => true,
+        '@PSR2' => true,
+        'no_unused_imports' => true,
+        'single_blank_line_before_namespace' => true,
+    ])
+    ->setFinder($finder)
+;
